requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
cassandra-driver==3.28.0
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.8.1
asyncio==3.4.3
tabulate==0.9.0
